Add uswds 0.12.1 to assets/vendor/
This is the quickest way at the moment, but if later on we decide to
use an asset pipeline like gulp, we might want to install uswds via
npm. I did actually give the npm install (not gulp) a try, and got
the sass @import working, but at the very least that still leaves out
javascript, and I suspect it also means the references to fonts and
images within uswds would break.