Create Jekyll plugin to generate pages from Contentful data
The jekyll-contentful-data-import gem maintained by Contentful does
what it says on the tin: it fetches data from Contentful's Content
Delivery API and puts it in a YAML file in the _data folder. Jekyll
loads that data file automatically, but everything else is up to us;
we get to tell Jekyll how to translate our data into a site.

I've done that by creating a generator (one of Jekyll's plugin types)
that maps the Sections and Pages of our Contentful content model to
what Jekyll calls Collections and Documents. By default, these render
into standard layouts (e.g. _layouts/section and _layouts/page), but
I've also made sure our generated collections can be configured via
the defaults hash defined in _config.yml (that's the main advantage
of a Jekyll::Document over a Jekyll::Page).

This commit also introduces the "custom" collection, which allows us
to associate a specific template with the Contentful ID of any Page or
Section, and thus extend the default templates with one-off
modifications or features. I'm not totally sure we'll need this in the
long run, but it could be useful during prototyping.

The last change in here renders the header menu from a menu entry
defined in Contentful. It's very basic right now (string matching
on the menu name), and I'm a bit undecided about whether it should
stay that way.

I'll be writing developer-oriented usage documentation that explains
how to render Contentful data in templates, how to customize
collections, and so forth, but you can get a head start by just
checking out the templates in this diff.