Add README, including Contentful-related developer usage docs
The README provides a quick explanation the project to anyone who
might find it on a GitHub, and more in-depth usage instructions to
developers working on the project.

I decided while writing it to add _data/contentful to the .gitignore,
instead of committing the YAML files it contains to master; that
change has since been committed to master. My thinking is that
committing changes to the data file in our topic branches could easily
become noisy and distracting, and even potentially introduce
cumbersome merge conflicts. The downside to _not_ committing it is
that every developer on the project currently needs Contentful API
keys to sync the updated data; especially once the data stabilizes
(i.e. by the time the new website launches), that might not be ideal.
At that point, our continuous deployment process could
push changes to the data file to a non-master branch, and we could
replace `jekyll contentful` with a command to simply pull changes from
that branch, no keys required.

The new README includes some inline TODO comments about what's still
left to add.