Configure Circle to download s3_website.jar during dependency phase
The `s3_website push` deploy command automatically downloads the .jar
file that does s3_website's heavy lifting, but the download is not
cached by Circle. This is a workaround to download the file during
the build's dependency phase, so that it'll be cached.

It's obviously a made somewhat fragile by being contigent on
s3_website's internals (https://github.com/laurilehmijoki/s3_website/blob/v2.15.1/bin/s3_website#L200-L245),
so I might open an issue on s3_website suggesting the addition of an
`s3_website download` or similar command that we could use instead.