Switch to jekyll-assets pipeline with npm-managed uswds
The primary motive for switching to a proper asset pipeline beyond
Jekyll's default support for Sass is being able to bundle USWDS from
source, so that we can override built-in variables including sizing,
spacing, and colors. (This is the recommended way to customize USWDS
per standards.usa.gov.) We will also of course take advantage of
jekyll-assets & Sprockets' other features, including asset hashing
for cache-busting and so on.

(NPM is the only package manager 18F officially supports for USWDS,
but that didn't prove to be a problem; getting Sprockets to play
nicely with node_modules mainly just required a couple of link_tree
directives to declare dependencies on the images and fonts.)

Probably the biggest downside to compiling USWDS ourselves is the
performance hit, especially because Sass's @import statement is
perhaps the slowest single part of a Sprockets process. I've tried to
minimize the cost by importing uswds in a separate file (uswds.scss),
and then using Sprocket's `require`, rather than import, to include
it in main.scss—using `require` means Sprockets will cache the
compiled version of uswds.scss. The caveat is that `require`, unlike
@import, doesn't share scope (variables and functions), so we may
need a higher-level variable file if we need to share some things
between our Sass and USWDS.

In general, as well, it's my understanding that we should aim to
minimize and localize imports of dependencies (like Bourbon
functions) to the specific scss files that need them, and then
(again) use Sprocket's `require` instead of @import—but we can do
some trial and error to see how that plays out in day-to-day use.

Finally, I'll add that I did get about halfway down the road of using
Webpack for this instead of jekll-assets/Sprockets. The Webpack
documentation leaves something to be desired, but Webpack itself
worked well, and certainly offered more granular control than
Sprockets (some might say too granular). That said, the sheer amount
of configuration required and the lack of plug-and-play integration
with a Ruby/Jekyll process (e.g. for referencing compiled/hashed
assets) eventually seemed like too high a hurdle to bother jumping
for this project.