Differentiate content blocks by type
Uses [jekyll-contentful-data-import][]'s "custom mappers" feature to
add a `type` property to the output yaml. At the moment, the default
`jekyll contentful` command doesn't know how to require custom
mappers, so we need to use custom `rake contentful` command. I opened
contentful/jekyll-contentful-data-import#22 to raise the possibility
of automatically requiring mappers, so we'll see where that lands,
but in the meantime we'll just have to switch to `rake contentful`.

I've also added a content_blocks include to handle the type-switching
and automatically render the corresponding partial from
_includes/content_block_types; it can be used like so:

```liquid
{% include content_blocks.html blocks=page.contentful.contentBlocks}
```

[jekyll-contentful-data-import]: https://github.com/contentful/jekyll-contentful-data-import