Style breadcrumbs as arrows with variable color graduation
Introduces $acc-section-colors to associate root sections (e.g.
Facilities, Services) with specific colors; if found, the color
associated with a given page's root section is now the background
color of the breadcrumbs bar (otherwise it falls back to a default).

At the moment, the last breadcrumb is the current page, and has a
transparent background; the background of each preceding breadcrumb
get progressively darker, with the 'Home' breadcrumb always being the
darkest (which is also the color of the left side of the bar).