Add generator to scaffold month-by-month calendar pages
A new CalendarGenerator creates pages for each month between the
first and last month found in the calendar data imported from Socrata
(including, right now, months that don't have any events). Events
that span the end of a month also appear on the next month's page.

The month pages render from a new "month" layout, while the main
/calendar/ page is now a Section in Contentful, with a custom
template in _custom (depending on the final design, it may end up
making more sense to simply add logic in the section layout, instead
of using an entirely custom layout, but that's TBD).

Breadcrumbs work on Calendar pages, with a minimal amount of
supporting logic now shared in a "Helpers" module between the
Contentful and Calendar generators.

Additional logic around filtering past events and so on remains TBD.