Render press releases
Each press release now renders as a page using the press_release
layout and blog-like URL that includes the date. They're added to a
press-releases collection corresponding to the Contentful section,
the index page of which  is rendered from _custom, much like the new
Calendar index. I still don't love _custom for this, but it works for
now.

Press release archive pages (by year) are not yet implemented but will
be added later.