Create homepage
This is the initial scaffold of the homepage that displays the 6 tiles based on the "HomepageTiles" menu.
It appears as though a lookup for items with spaces in their names does not work, so I created the camel-cased version in Contentful as a workaround.

There is still more work to do, such as splitting words onto separate lines in the link text and cleaning up the border handling, but this will be sufficient for our Alpha version.

We should probably add some logic to breadcrumbs handling to remove it from the homepage, but I added embedded CSS in `index.html` as a stop-gap to hide it.