Add Footer
This initial scaffold of the footer includes a combination of hard-coded links as well as Contentful-driven menu content.

There is some CSS refactoring as well, most notably a `link-color-states` mixin that accepts a color variable and appends it to `&:hover, &:visited, &:active` pseudo classes

Since we do not have all SVG icons yet I used a combination of placeholder images and inline text, and created a Github issue for the icons to be updated once they are completed in design.