Implement responsive image gallery and viewer (PhotoSwipe)
Gallery blocks now render as responsive grid of thumbnails cropped to
a 3:2 aspect ratio, and display in full (also responsive) sizes in
PhotoSwipe (photoswipe.com).

I picked PhotoSwipe from GitHub, where (by my estimation) it's the
most popular JS image viewer by far (12k+ stars). It has solid
documentation and an active community, and fairly extensive options
for customizing both the appearance and behavior. It's completely
markup-agnostic, which means we can use Schema.org's ImageGallery
markup.

The new JS file here is based on PhotoSwipe's example code, but
it doesn't bother with IE8 compatability and has a couple notable
differences:

* The responsive image sizing uses Contentful's Images API. Since
  PhotoSwipe needs the dimensions among its options, we calculate
  them based on the original image's dimensions, which are output by
  the new GalleryBlockMapper and rendered as data- attributes.

* It uses Contentful IDs for the gallery ID and picture ID that
  PhotoSwipe sets in the URL hash, which means that permalinks should
  still work regardless of the order of gallery blocks and images
  within them.

The responsive thumbnail grid uses flexbox and shows either 2 or 3
images per row depending on the viewport. I spent a lot of time
trying different flexbox options; my original hope was to crop the
thumbnails in CSS so that we could enable PhotoSwipe's nice
transition animation using the uncropped thumbnail. In theory, we'd
want to say that a thumbnail is 33% wide and has a height of 2/3rds
of that, hiding the overflow, but percentage heights are sort of a
joke; CSS ignores them unless the parent has a fixed height.

Thus flexbox loves to render each row at the height of the tallest
thumbnail. I *could* set a best-guess max-height in pixels, but that
changes the aspect ratio of the thumbnails as the viewport scales,
and the transition animation looks almost as janky as starting with a
pre-cropped thumbnail. I eventually just gave up, cropped the
thumbnails via the Contentful API, and disabled the animation.

The flexbox styles are fairly straightforward, the only maybe-unusual
part is the nth-of-type declarations. Flexbox's justify-content rule
doesn't work well if rows don't have the same number of items, so
those are needed to have edge-to-edge rows with the last row aligned
left.

The only additional gallery feature that I think we might want to add
would be limiting the number of thumbnails displayed, and having a
"View All" button open PhotoSwipe if there are more than the limit.
Whether that makes sense depends a lot on how many photos there end
up being, so we can wait and add it later.