Update floor plan styles to render a setups list as flexbox grid
The definition-list approach to marking up the floor plan setups
ended up being not very semantic, so this now simply expects an
unordered list of class `acc-floor-plan-setups`. Kramdown will render
Markdown inside each item given a `markdown="1"` attribute on the
`<li>` tag, so the editor's version is still pretty clean. For
example:

```
<ul class="acc-floor-plan-setups"><li markdown="1"
class="acc-banquet">

66" rounds: 2,000 guests \\
72" rounds: 1,430 guests

</li></ul>
```

I suspect we may want to abstract the flexbox-list classes (or create
a mixin) so that they can be used elsewhere (e.g. specifications),
but for now they're just added to the setup classes.