Implement responsive side navigation and add sprockets-commoner
Updates the sidenav component added in a previous commit to hide it
in a sliding panel on devices with smaller viewports (as determined
by the $large-screen breakpoint, which is where the 1/4 + 3/4 grid
breaks). The styling of the panel is heavily borrowed from the USWDS
equivalent, which is still used at the moment for the main nav.

This commit also includes the introduction of [sprockets-commoner], a
Sprockets extension maintained by Shopify that adds babel-powered
postprocessing of of CommonJS-style `require` statements. The sidenav
and folders components are now require'd in main.js and listen for
events through a main event delegate (an instance of [dom-delegate]
bound to `document`). Though they're not changed by this commit, the
Photoswipe and floor_plan.js will be able to adopt this pattern as
well.

The only gotcha I encountered was the need to narrow the files for
which sprockets-commoner tries to resolve `require`; by default, it
will resolve any JS file in the Sprockets paths, which ended up
causing problems when bundling the Tether lib. The solution was just
configuring commoner (via _plugins/hooks) to only process files in
_assets/javascripts and the specific npm modules that we want to
`require` (e.g. dom-delegate).

[sprockets-commoner]: https://github.com/Shopify/sprockets-commoner
[dom-delegate]: https://github.com/ftlabs/ftdomdelegate