Homepage Tile Makeover, Breadcrumb responsive fix, footer responsive fix, removes back to top
Changes the homepage tiles to a purple on hover that corresponds with
the banner image.  I think the color looks better than just going to
darker gray.  I fixed up the responsive spacing of the buttons on
smaller screens as well.

I added CSS when screen size is under 600px to hide all breadcrumbs
except the current page and it's parent, if it has one.  The font size
scales down to 14px to help the potential 2 tabs fit better.

Added footer responsive css to make the quick nav lists line up better.

Removed the back to top link.