Match button styles to design spec and enable button grid layouts
The primay button style now expects an H3 title and an optional
description rendered (usually with paragraph tags) inside of a div.
Section pages now feature a two-across flexbox-based grid of primary
buttons, while the homepage tiles have been swapped out for a version
of the same grid with three buttons across.

The one outstanding issue with this implementation is that the
homepage button descriptions will only work, at the moment, if the
menu links to the Page object, rather than the Section object. We can
work around that for now, but ideally we'll have something like
url_for (or something that replaces url_for) so that we can use
Contentful IDs to get page attributes other than just `url`.