Support option to render gallery as full-width images
Instead of thumbnails that open an image viewer. This lets content
editors use gallery blocks when they only need to add one or two
images to a page, which is preferable to using a text block because
it means the asset can be replaced in Contentful without needing to
also edit the text blocks that reference it.