Configure s3_website Cloudfront options
Includes Cloudfront settings for creating the distributions and
caching policies to set on the files when deploying. s3_website 3.0.0
notably includes a slightly modified syntax for configuring
Cloudfront, as it's now using AWS's Ruby SDK.

In addition to the usual advantages of a CDN, Cloudfront integrates
Amazon Certificate Manager to issue (and rotate) SSL certs.