Add a simple rake task to trigger nightly site rebuilds via Heroku
Intended as the simplest possible way to trigger a rebuild of the
both sites each night, which needs to happen to ensure the Calendar
stays up-to-date even if updates aren't made via GitHub or Contentful
in the current month.

It's obviously a somewhat naive solution, but the risks of failure
are fairly low, and low-cost even if they do happen. As the City
invests resources in building out more static-site infrastructure,
it's possible (even probable) that this will be replaced by something
a little more robust (e.g. with error handling, and perhaps using
Lambda w/ Cloudwatch events instead of Heroku Scheduler).