Update README with workflow to change Contentful content model
Outlined the basic procedure to spin up dev spaces from production
data, make content model changes, and then import those changes back
to the production spaces. The workflow relies on Contentful's export
and import tools, which have replaced the contentful-space-sync tool
that was previously referenced in the README.

Some of the documentation related to the import tool suggests that it
only works on "new and empty spaces." I'm not sure if that applies
when you're importing content, or if it's just a recommendation, but I
used the tool to import --content-model-only changes to existing
spaces and it seemed to work just fine.