Add Facebook campaign analytics
This includes 2 levels of FB campaign tracking - a page view and what I believe is a paid campaign referral tracking that is isolated to the "Book an Event" page. I was not able to confirm the efficacy of the tracking snippet due to lack of access to the associated FB ads campaign. I am assuming that I can daisy-chain the `PageView` and `Lead` functions within a single `<script>` tag but without access to the campaign itself I cannot confirm this.