Rename "Calendar" as "Events"
The primary goal was for the "Calendar" page/section to be renamed "Events" on the user-facing parts of the site. The change could have been isolated more than it is here, only making the substitution where the calendar builder is looking for the "Events" content item from Contentful, but then we would have been left with routes that say "/calendar" for a top-level page titled "Events", and the code that supports this feature would have been semantically inconsistent, and therefore confusing. This resulted in my decision to replace all references to "Calendar". The one place this gets weird is in the nesting of individual "events" under the "events" parent list (instead of "events" being children under "calendar" but) but I felt like that was a more reasonable approach than turning "events" into "calendar" at some point in the import/build chain.

Since these code changes are tied to Contentful changes this code must be coordinated with the publishing of the changes to the "Events" Page and Section in Contentful.