Add support for homepage press statement
This allows a content manager to quickly add a text block to the homepage in case an urgent statement needs to be published. This was prompted by the need to issue a statement addressing event security amidst a serial bombing event in Austin.

See the _Homepage Guide_ section of the site Guides to learn more about usage.