Add disclaimer to events page
I initially tried doing this via Contentful, but ran into complexity in the site build process since all of the events/calendar pages are currently generated purely with Socrata data. To add a relationship between the contentful data and the events page within the build process seemed like far more effort than it was worth to make a couple pieces of static text that are not likely to change more than 2x/year editable via Contentful.