Embed Curator.io social media widgets on homepages
Each site has a corresponding widget configured in ACC's Curator.io account. Widget IDs are set as site config vars.

One thing to look out for is that the expected target element for the javascript DOM injection is configured on the Curator.io side, so it's critical that the ACC and PEC homepage widgets have this value set to "curator-feed". It looked like the PEC feed was auto-configured to "curator-feed-pec" which caused some confusion as I was trying to debug why PEC's feed wasn't loading.

The embed code I used is the one generated from the widget creator tool. There are separate docs with other embed methods, including one that uses jQuery and provides a more digestable JSON format for setting widget params like type and ID, but I was getting CORS errors locally and live. I didn't think it was worth diving too much into since the original implementation was a Curator-generated snippet.

As noted in the Homepage Guide in the site guides, each site (prod _and_ staging) has its own widget, so there are 4 in total. The ACC Staging widget is completely independent from the ACC Production widget. Similarly with PEC. This means that the ACC Staging widget will only appear on ACC Production, and if you want to make the ACC production widget look the same then you need to manually apply the same changes to that widget and publish them in Curator.io. It is not possible to "promote" the staging widget to production, at least as far as I have been able to tell. The other thing to note about Curator is that each source:widget pairing requires a dedicated source. This means that if I want to show the ACCD Instgram source on the ACC Staging and ACC Production widgets then I must use 2 of the available Sources on our plan. It is not possible to just say "syndicate this 1 source to multiple feeds". This might create some content scaling issues in the future.