Add 'plan' section templates to support the new IA
The RFP form and Floor Plans have been moved to the Plan section. A previous change introduced floor plan templates for several different parent sections, which I have removed as I could not determine the reason for doing it and cannot identify a negative consequence of removing them in favor of a single floor plan template within the _plan/ templates directory.

This also includes a renaming of the RFP page.