Add 'Page not found' title to 404 for Google Analytics
During research around the recent production errors I was trying to find analytics around our 404 page. I infer that pages with the title 'Austin Convention Center' (and no additional text) are errors, unless the page path is '/' (the homepage). Adding this to the page title for the 404 will make it much easier to tell with certainty that these are in fact indicative of Page Not Found errors rather than a missing title in a template (which I do not believe to be the case).