Change value of ACC "Calendar Location" site variable
Long-story short, improperly formatted data was introduced to Socrata causing the build process to fail for a number of days. This appears to have been caused by Terri delivering an events spreadsheet to Chris that had "Austin Convention Center Events" instead of "Austin Convention Center" in the Location column for ACC events.