Add Google Analytics event tracking
This implements tracking of the following events:

Photo Carousel

 - When the Previous button is clicked
 - When the Next button is clicked
 - When the thumbnail is clicked to render the full-screen photo
 - When one of the pagination dots is clicked
 - When the carousel slides to a new photo
 - When the carousel is loaded

Homepage Hero Buttons

- When a button in the hero menu is clicked

Form Submission

- When a submit button is clicked
- When a form actually submits

External Links

- When a sidebar PDF button is clicked
- When a sidebar external link button is clicked
- When a text link within a content block is clicked (the .acc-inline-external-link class must be applied via markdown in the Contentful content item)

---

Notes & Caveats

The photo carousel plugin has nice documentation on all the events that can be tracked.
https://sorgalla.com/jcarousel/docs/reference/events.html

Only the carousel has click tracking, as I confirmed with Leslie that it is the only photo gallery type in-use on the site (and it appears the original tile gallery may have been deprecated altogether).

Though GA has documentation for tracking external links differently I found that using the same script for all links functioned locally and live, so I removed the outbound-specific function which included specification of an additional post-callback redirect URL.

The additional weight of jQuery is ~70k which I felt was a reasonable tradeoff for all the things we can use it for in the future, and since there's a chance a visitor might already have the CDN version cached on their device. That contributed to my decision to reference the CDN-hosted version rather than a local one.