Add deployment configuration for staging on Circle 2.0
ONLY STAGING!!! DOES NOT TOUCH MASTER BRANCH!!

The one thing that CircleCI didn't auto-translate was the deployment configuration. They recommend using _workflows_ for deployment rules but for our use case I think it's okay, at least for now, to just use a branch filter on the deployment step. I initially tried defiining a new 'deploy' configuration at the same level as 'build' in `.circleci/config.yml` but eventually found this approach to be the working solution. It still may be worth looking into workflows in the future but, again, I'm not sure whether it's worth it for what we're doing.

To add master just add it above/below the staging value in the deploy filter.