Add deployment configurations on Circle 2.0
This migrates our build and deploy pipeline to CircleCI 2.0 ahead of their 1.0 deprecation at the end of this month.

The one thing that CircleCI didn't auto-translate was the deployment configuration. They recommend using _workflows_ for deployment rules but for our use case I think it's okay, at least for now, to just use a branch filter on the deployment step. I initially tried defiining a new 'deploy' configuration at the same level as 'build' in `.circleci/config.yml` but eventually found this approach to be the working solution. It still may be worth looking into workflows in the future but, again, I'm not sure whether it's worth it for what we're doing.

I'm still a little paranoid about the build time stuff, so as an extra precaution I've prevented feature branches from triggering builds on CircleCI. I don't think this will have any negative consequences on testing since all of the dev work ends up on Staging at some point anyway before being PRed into Master.