Scaffold the Learn More menu
This lays the initial foundation for the new Learn More Menu, which consists of 3 Learn More Links that direct a visitor to either a Page on the site or to an External Link.

This does not including styling work yet. My intention is to first get all of the Contentful => Jekyll functionality working and to solidify the content management experience before moving onto the purely visual tasks around layout, typography, and color.