wip scaffold Page Banner
Includes global layout updates to accommodate full-screen items nested at the Content Block level in the template. I decided to take this approach to the Page Banner, rather than leaving it hard-coded at the Page template-level, to allow for easier maintenance of that part of the page for editors. This will be a point of discussion when the initial implementation is presented to key stakeholders.