Implement new megamenu-based navigation
- Removed the toggle carrots on the dropdown items
- Hard-coded Events link since it has a visual deviation with the left pipe and since it shouldn't really change
- implemented initial color fade pattern as a mixin and applied to the subnav sections

Got caught early on with USWDS's sensitivity to/reliance on ID names for the subnav menu toggle, so that's where the `{% case %}` statement comes into play, to translate the indexed order of the nav item (1, 2, 3, etc...) to a semantic name (one, two, three, etc...) in order to append it to the `basic-mega-nav-section-IDENTIFIER` ID declaration.

I had to do a less-than-ideal override of the default column widths because USWDS only gracefully supports a 4-column layout there and won't accept overrides with `usa-width-one-third`. The code is less than ideal but I think we accomplish everything we need to from a design, compatibility, and accessibility standpoint so I'm not going to bang my head against a wall and keep on billing for a prettier solution that only I or future maintainers will ever notice.

The fixed constraints/assumptions on the megamenu from USWDS, and the layout hack mentioned above, made incorporation of the social nav icons less plug-and-play than I had hoped. I decided, at least for this commit/batch of work, to just leave them out and address again later in a follow-up with the team. The social links will remain accessible in the site footer.