Track clicks on RFP external link instead of deprecated RFP Form submissions
The decision was made to use Ungerboeck's white-label form creator to create temporary RFP forms until we have a native solution for website-to-USI RFP submission in place. Accordingly, the RFP text block in contentful has been updated to provide a link to the form instead of a solictation to email the sales team. This analytics change removes the old form submission tracking and replaces it with click tracking on the new RFP link, which is used on both the ACC and PEC sites.