refactor nav.scss to be mobile first
The goal here was to remove media queries that refered to both mobile and desktop - which gets confusing. Now it is (mostly) mobile first, with the mq being desktop.

I say mostly, because I could have spent a few more hours on this refactor, but thought my time was best used moving on to other "remove uswds" things.