Create KML assets directory
I tried to create this in the root `_assets/` level but couldn't get the Jekyll asset path to find it. I looked up some configuration settings that were relevant but they seemed to be only accessible with more recent versions of some Jekyll-Assets dependencies (Like Sprockets 4+ in particular) so for now I'm just tucking them inside of `images/`, which allows them to be targeted with the `asset_path` helper.