Show upcoming calendar events on home page
The decision was made to handle the events in JavaScript instead of Jekyll, because doing it in Jekyll would require the site to be rebuilt to update the current events.

While it's possible that more event handling could have been done in Jekyll or Ruby, JavaScript is necessary regardless, so I opted to just use what existed.