Temporarily disable custom Google Maps markers
Deferring to default markers until we have our custom ones designed and added to the assets dir.

Note that even though the marker variables are commented out Jekyll still tries to interpret them and will throw a build error if a corresponding asset is not found. That's why I replaced the 'marker-n' file names with the coa-stamp.