Reconfigure Mega Feature to behave like Side-By-Side feature
After several attempts at wrangling the balance between the text length, image ratio, and device width we decided to simplify this feature and treat it like a full-width side-by-side menu. This involves a change to the Content Model to make the Mega Feature similar to the side-by-side in that it validates the presence of 2 Items (which are References to either a Text Block or Gallery Image) and displays them in the order in which they are listed in Contentful.