Clean up Analytics event tracking and add support for Photoswipe
The previous tracking was set up to track events from the JCarousel implementation last worked on during the previous version of the site. Now that Mike's work has re-introduced Photoswipe I have updated the tracking accordingly. Due to limitations with the Photoswipe API and listener support I decided to abandon forward/next tracking support in favor of just getting each instance of a full-size photo load, which can be triggered by expanding a thumbnail from the grid view, by clicking the previous or next buttons, and by swiping left or right. If more granularity is desired in the future we can certainly add Previous/Next click events, and possibly swipes as well but I'm less optimistic about that.

I also realized that some references to the sidebar-external-link were still in the site template (particularly in the Menu component), so I've removed those so now only the "External Link Click" will be present in the data going forward.