Disable Google Analytics in development
Event tracking was pinging the production GA property even though the site was being served locally. This prevents that from happening. For testing GA stuff in the future, developers can simply temporarily remove the condition in `_includes/analytics/google_analytics.html` and serve the _staging_ version of the site(s) locally (so the analytics will ping the Staging GA properties instead of the Production ones).