Deprecate the Press Release Contact Block
With the new redesign, the only place on the site using the Contact Block was the Press Release, and all of the Press Releases just list Terri McBride as the contact person. After discussing with Leslie Barrientos we decided to just incorporate Terri's contact information in the layout itself, rather than making it a thing that needs to be configured and managed on the Contentful side every time.

As this is essentialy site "data", similar to the facility addresses or social profile URLs, I decided to make a press_contact variable scope in the base `_config.yml`. This felt more appropriate that adding these values directly into the HTML template.