Move Calendar of Events source from Socrata to Contentful
As it turns out, data in USI and in Socrata is often incorrect, which meant that problems exposed as symptoms on the website (bad calendar data) had to be mitigated with a whole chain of dependencies (USI data, the data export from USI, the transposition of that into a spreadsheet, the upload of that spreadsheet to Socrata, and then finally the rebuild trigger on the website to rebuild the calendar according to the new Socrata data).

This new process takes USI and Socrata entirely out of the equation, favoring instead a much more tightly-controlled approach by which content editors manually add and publish new Events directly in Contentful. A new Content Type has been added to both Spaces, and that data is being used to generate the calendar. In this new process, if a problem is discovered with the Calendar of Events, anyone who has Publishing privileges via Contentful (or via the Web Admin Dashboard) can quickly fix it and redeploy the site(s).

In order to reuse the logic that was originally built to handle the page-by-page calendar buildout the events data is still stored as a YAML file that is parsed by the calendar generator, the only difference now is that instead of generating one file for all events through a standalone rake task, a calendar file generator is called as soon as each space's Contentful Import has completed. I couldn't figure how to access the `site.***` object in the Rakefile (I think it's not possible?), so instead of getting an array of events that way I'm instead kicking off an additional API call to get the Space's Events, roll them into an array, and then dump that into a space-specific YAML file in `_data/events`.