Remove year/month/day paths from Press Release permalink
The marketing team reported a problem with Press Release URL handling which resulted in all Press Release URLs being generated with the deploy date set as the datestamp, rather than the Press Release date attribute itself. I dug through the codebase trying to identify where that was getting lost (the document "date" value correctly persists throughout the process chain but seems to be ignored at URL generation), and also trying to identify if this is somehow something new or if this has always been an issue that we just never noticed. I'm confident that this behavior was correct when we first launched the site, and even went so far as to downgrade my own local Jekyll to see if there was some other change to stack dependencies that would have somehow modified the process by which Jekyll extracts a date from a Page and applies it to the permalink structure.

For now the quick fix is to just remove that datestamp all together, so
...center.com/press-room/YYYY/MM/DD/TITLE just becomes
...center.com/press-room/TITLE

Old links to press releases will no longer work as there is no concept of "Redirect URL" applied to the Press Release Content Item in Contentful and I think it would be more trouble than it's worth to support that for something as infrequently-viewed as Press Releases.