Add support for Contentful-driven press contact
This provides a balance of greater styling with easier ability for content managers to provide contact information relevant to each particular press release. This is achieved by adding a "Contact" text field to the Press Release Content Model, and then rendering that in a div with particular styling rules applied.