Replace s3_website gem with CCI-native AWS-CLI tool
On May 15 we started getting build errors related to the S3 website push. After rolling code back to about a week old version that had been built dozens of times I was able to replicate the error on staging, leading me to investigate CircleCI and AWS. What I discovered was the gem we're using to coordinate between the two, s3_website, has been out of support for almost a year and a half, and that other devs had experienced issues as a result of non-maintenance.

That's when I discovered CircleCI's _orbs_, among which is a CCI-developed orb that installs the AWS CLI into the build environment, allowing us to sync a specified directory using the CLI itself, from within the CCI Shell. Nice!

Once that was running, however, I realized that the s3_website gem was also responsible for invalidating the CloudFront cache after upload. Fortunately I was able to google stackoverflow my way to the solution presented here, which requires us to manually permit CLI interaction with CloudFront (disabled by default), and run the invalidation command from the same line as the sync command (in Rakefile).

The changes will need to be reflected across site docs.